Update git-setup.sh
We have modified the check_exit_status function to write messages to a log file in addition to printing them to the console. With these changes each step's success or failure will be logged to both the console and the specified log file (/var/log/script_execution.log). This provides a record of what happened during the script execution, which is useful for debugging and auditing.